+++ figpack-figure/src/figpack-main-plugin/views/TimeseriesGraph/drawFunction.ts
@@ -98,6 +98,7 @@ export const createDraw = ({
             visibleTimepoints,
             canvasWidth - margins.left - margins.right,
             s.downsampleFactors,
+            o.exporting ? 2 : undefined,
           );
 
           let uniformData;